[wip] Hand impl our 'to bytes' trait
It's called 'Table', and this is not a good name.

This gets us to a point where (fingers crossed) I would be able to
compile a gdef table.

It's all pretty untested, though, so what I would like to do *now* is to
write a little binary that reads gdef, deserializes it, inspects it, and
re-serializes it. Does this work?