[codegen] support 'use' statements
This seemed necessary when I was adding types to offsets, but I'm now
less sure about how that should work, and this may end up not being
necessary? But we don't want to rewrite it if it *is* necessary,
so let's get it in here for now, and we can remove it later if it is
unused. :shrug: