[codegen] move to include! macro for generated code
This is much simplier from a hygene perspective, because we aren't
creating a bunch of semi-secret modules that we don't tell anyone about.

It also means that the user can easily add code that lives in the same
module as what is generated, which makes life simpler in a bunch of
ways.