[codegen] Add TableRef type aliases
okay so we ran into a funny problem: the concrete type of a non-format table
is `TableRef<'a, MyTableMarker>`, but the concrete type of a format type is,
like, `AnchorTable<'a>`. This sucked, because we cant consistently know what
kind of identifier we should be expecting to return from an offset.

Type aliases to the rescue! if we make `type MyTable<'a> = TableRef<'a, MyTableMarker>` then we get consistency, and this seems reasonable.

This also fixes up some stuff with generating only the docs we need, and
hiding various other things.

It is a signpost on the path to maybe getting rid of our marker types
altogether?