[codegen] Support records that contain references
This is... a compromise. It is nice to think of records as being
basically tuples, but there are a number of records in the spec that are
variably sized.

With this patch, records now follow one of two paths: either they are
'simple' records (which can be cast from bytes) or they are not, in
which case still have their fields init'd at creation time (unlike
tables, which init them on access) but which can contain references to
font data; for instance they can contain a computed array, where the
individual items are init'd when they are accessed.

This definitely makes things more complicated, but it also lets us
continue to keep more stuff in codegen, and that feels like a win?