[WIP] First pass at a custom pretty-printing impl
This implements the start of a custom approach to pretty printing, that
will provide greater control. In the future, I imagine this utility
expanding so that it accepts arguments, and the caller can do things
like choose the level of verbosity, avoid printing the contents of
arrays, and print items at specific paths, sort of like the jq command
line tool.