[codegen] Support phantom generics in generated codegen
This is a pattern we use in a number of places, where some table has an
offset, and the type of that offset varies depending on some type
identifier.

This patch is the first part of handling this: we can now specify that a
table has a phantom generic, in which case we will add an appropriate
PhantomData param.

This was only slightly annoying, actually, and this is now being used to
replace the TypedLayout, TypedExtension, and TypedLayoutList types,
which were previously hand-written.

The next step is going to be to generate wrapper enums that contain all
of the variants allowed for these types, and then to also handle this
appropriately on the compilation side.