[codegen] Add 'test' codegen input
This is an input file to the codegen system that is intended to
improve the developer experience when making modifications to the
codegen code itself. By having a single, simplified input file, we can
selectively rebuild only that file during development, avoiding the
situation where errors in generated code cause a large cascade of
errors.

This file is currently quite barebones; my intention is to add new
tables to it that exercise various codegen paths, as issues arise.