Add types to represent typed arrays of offsets
This was a long-standing papercut: in codegen we generate typed getters
for offsets, but for arrays of offsets this was limited to generating an
iterator.

This was a slightly annoying API, since an iterator cannot be retained,
and does not provide random access.

This patch addresses this by adding a pair of concrete types that can
wrap arrays of offsets and nullable offsets, and can dynamically resolve
those offsets as they are accessed.