[codegen] Make #[count] not accept expressions
This introduces a new, more constrained grammar for this attribute,
supporting two forms.

In the simple form, it accepts a single argument, which may be either a
`$field`, a literal integer (`52`) or the special token `..`, which
means "all remaining data", and is only valid for the last field in a
table.

In the non-simple form, it accepts a function like syntax, where the
'function' is one of a small set of predefined options. These functions
expect some set number of pre-defined arguments, which may be either
`$field`s or literals.

supported 'functions' are currently `subtract`, `half`,
`delta_value_count` and `map_data_size`.