[codegen] Make #[available] not accept expressions
Further work on making us schema-ready. Previously this attribute
accepted any expression that evaluated to something which implemented
the 'Comparable' trait, for the appropriate version type. In practice,
this was so that we could provide both literal integers as well as named
constants, such as `MajorMinor::VERSION_1_1`.

The schema will not support the concept of an arbitrary expression,
however, and so this needs a more constrained representation.

With this patch, we no longer support the use of named constants in this
attribute; instead the argument must be either a single integer literal,
or a comma-separated major/minor pair of literal integers, which we will
convert to the appropriate type when comparing.