Small changes to fixed point types and CharIter (#199)
* Small changes to fixed point types

This adds `to_bits` and `from_bits` conversions to the fixed point types. We can work toward better typing in punchcut, but these feel reasonable and will alleviate some of the pain of roundtripping through bytes in meantime. This also adds a conversion from `Fixed` to `F2Dot14`.

Additionally, derives `Clone` for `name::CharIter` in the spirit of making iterators clone where possible. I ran into this when working on localized string support in the new font metadata library.

* move to_f2dot14 comment up to docs