[codegen] Add #[derive_traits(..)] attribute
This was an old todo that I hit again when working on gvar; sometimes we
want some particular table/record to implement a specific trait, on a
case-by-case basis. This adds support for a new table-level attribute
where you can provide a list of traits that will be derived for the
type.

While getting this to work I ran into a few existing cases where we were
not deriving traits that we could be, so I added those as well (for
instance, we will now derive Ord/PartialOrd for generated flags as well
as BigEndian<T>).