Add TableType
This is a way for us to keep track of the type of a table after it has
been serialized, which is necesary if we later want to go and modify it,
such as for promotion or extension.

This overlaps somewhat with the `name` field added for graphviz, and
they could possibly be combined? I need to think about this a bit more.

fixup lookuptype