Add 'serde' feature to read/write-fonts
This is intended to be used to serialize test cases in write-fonts; it
is also added in read-fonts since write-fonts reuses flags and enums
declared in write-fonts.

One funny thing: (I think) this hit a serde bug around using
non-literals to provide default values for const generics, which we do
in the OffsetMarker types in write-fonts.