Cleanup cmap test assertions
I'm sorry Rod, but this patch is going to make you mad.

Noticed this when reviewing old PRs; I thought it shouldn't be necessary
to make everything a vec, since there should be PartialEq impls for
comparing BigEndian types to their native values, and that should mean
you can then compare slices directly, since that's provided by std.

And this works! With one small, infuriating caveat:

*We can impl PartialEq<T> for BigEndian<T>, but not the other way
around*.

What this means is that you can write, `be_val == ne_val` but not
`ne_val == be_val`. This is related to the orphan rules
(https://doc.rust-lang.org/error_codes/E0210.html) and the tl;dr there
is that we can't impl a foreign trait (PartialEq) on a foreign type (T).
So we are stuck with this weird asymetry.

That said, I thought it would be useful to PR this patch so that other
folks are aware of this fact, and can avoid confusion in the future.