[skrifa] add owned version of LocationRef
Adds an owned version of the `LocationRef` type called `Location` for storing computed normalized coordinates.

Will be used in a followup PR computing normalized locations.

Also adds a `SmallArray` utility type which is really just a limited version of the type available in the `smallvec` crate. This is used as the backing store for `Location` and will also be used for other things in skrifa like computing avar2 adjustments and filtering variation settings.