check that p0,p1,p2 are collinear as well as equidistant before dropping p1
Before we were simply testing that the euclidean distance between p1 and p0 was the same as the one between p2 and p1, without actually enforcing that they were on the same line.
Here we check that the two vectors p1p0 and p2p1's x/y components are the same.
FontTools also does the same, see:
https://github.com/fonttools/fonttools/blob/66ce7c0e0c0ffdd11b2b5011323305f166ee3b85/Lib/fontTools/ttLib/tables/_g_l_y_f.py#L1599-L1601