[font-types] Serde Tag as str when target is text
That is, if the target format is human-readable we will serialize Tag as
a string, and otherwise we will serialize it as raw bytes.

Does this make sense? mmaaaaybe? It's definitely nice to serialize as a
string. Whether we even need to think about supporting non-utf8 tags is
unclear to me; they may exist in fonts in the wild, but they shouldn't
be used in new fonts.