[write-fonts] Add 'PendingVariationIndex' table
This is intended as a sentinal, which can be assigned a unique
ID during compilation, and then remapped to the final VariationIndex
values once the ItemVariationStore has been compiled.

This type will panic if we attempt to serialize it, which will ensure
that we never accidentally write temporary values to our VariationIndex
tables.

Although this is not part of the spec I am generating this in codegen
since it's much easier than trying to do it manually; making this work
required adding a new attribute, `#[write_fonts_only]`, to communicate
to codegen that we should not generate any parsing code for this type
(since it is not a real type that should ever exist in an actual font
file).