Add basic custom types for the 'meta' table
This adds types and the required impls that can represent the expected
contents of the 'meta' table.

This has to be handwritten, because the 'meta' table contains different
payloads depending on the included tag for each record.

I think having these as real types will provide a significantly better
API though, and avoid us writing invalid data.

The actual implementations for these types are very spartan; eventually
I imagine having parsing & constructing code for the ScriptLangTag
struct that ensures it conforms to the spec.