[codegen] Better handling of computed lengths
This fixes an issue where we were incorrectly truncating the all
computed lengths to be u16s, even if the value would be stored as a u32.

We now keep computed values as usize and then generate a conversion for
the appropriate type only at the callsite, instead of blindly converting
to a u16 earlier on.