[skrifa] match FT for CFF to autohint outline conversion  (#1568)
Our CFF path to TrueType-style outline conversion that is used to feed the autohinter was generating a slightly different point configuration leading to tiny autohinting differences. This aims to match FreeType by omitting the last on curve of a contour if it matches the starting position.

Tested against Material Icons and Cantarell VF which now match for all autohinting modes.